updated build scripts, AnimationFrameCharset typings

+++ Makefile
@@ -2,14 +2,18 @@
 SUFFIXES =
 .SUFFIXES: .c .cpp .h .hpp .rc .res .inl .o .d .asm
 
+
 #==============================================================================
 MAKEFLAGS += --no-print-directory
 #==============================================================================
 # Build platform
-PLATFORM?=windows
+PLATFORM?=linux
 # Build description (Primarily uses Debug/Release)
 BUILD?=Release
 _BUILDL := $(shell echo $(BUILD) | tr A-Z a-z)
+ifeq ($(BUILD),Tests)
+	_BUILDL := release
+endif
 
 # The sub-folder containing the target source files
 SRC_TARGET?=
@@ -26,11 +30,16 @@ DUMP_ASSEMBLY?=false
 # Clean output?
 CLEAN_OUTPUT?=true
 
+# If dll, build as a static library?
+BUILD_STATIC?=false
+
 # Platform specific environment variables
 -include env/.all.mk
 -include env/.$(_BUILDL).mk
 -include env/$(PLATFORM).all.mk
 -include env/$(PLATFORM).$(_BUILDL).mk
+
+# Target specific variables
 ifneq ($(SRC_TARGET),)
 -include env/$(SRC_TARGET)/.all.mk
 -include env/$(SRC_TARGET)/.$(_BUILDL).mk
@@ -81,20 +90,72 @@ SOURCE_FILES := $(patsubst $(SRC_DIR)/%,%,$(shell find $(SRC_DIR) -name '*.cpp' 
 PROJECT_DIRS := $(patsubst $(SRC_DIR)/%,%,$(shell find $(SRC_DIR) -mindepth 1 -maxdepth 99 -type d))
 
 # Add prefixes to the above variables
-_LIB_DIRS := $(LIB_DIR:%=-L%/) $(LIB_DIRS:%=-L%)
 _INCLUDE_DIRS := $(patsubst %,-I%,$(SRC_DIR)/ $(LIB_DIR)/ $(INCLUDE_DIRS))
 
 _BUILD_MACROS := $(BUILD_MACROS:%=-D%)
 _LINK_LIBRARIES := $(LINK_LIBRARIES:%=-l%)
 
+#==============================================================================
+# Unit Testing
+TEST_DIR :=
+ifeq ($(BUILD),Tests)
+	TEST_DIR := test
+	SOURCE_FILES := $(SOURCE_FILES:Main.cpp=)
+	SOURCE_FILES := $(patsubst $(TEST_DIR)/%,.$(TEST_DIR)/%,$(shell find $(TEST_DIR) -name '*.cpp' -o -name '*.c' -o -name '*.cc' -o -name '*.rc')) $(SOURCE_FILES)
+	_INCLUDE_DIRS := $(patsubst %,-I%,$(TEST_DIR)/) $(_INCLUDE_DIRS)
+	PROJECT_DIRS := .$(TEST_DIR) $(PROJECT_DIRS)
+	BUILD_FLAGS := $(BUILD_FLAGS:-mwindows=)
+endif
+
+#==============================================================================
+# Linux Specific
+PRODUCTION_LINUX_ICON?=icon
+
+# The full working directory
+ifeq ($(PLATFORM),linux)
+	_LINUX_GREP_CWD := $(shell echo $(CURDIR) | sed 's/\//\\\//g')
+endif
+
+#==============================================================================
+# MacOS Specific
+PRODUCTION_MACOS_ICON?=icon
+PRODUCTION_MACOS_BUNDLE_COMPANY?=developer
+PRODUCTION_MACOS_BUNDLE_DISPLAY_NAME?=App
+PRODUCTION_MACOS_BUNDLE_NAME?=App
+PRODUCTION_MACOS_MAKE_DMG?=true
+PRODUCTION_MACOS_BACKGROUND?=dmg-background
+
+ifeq ($(PLATFORM),osx)
+	PRODUCTION_MACOS_BUNDLE_COMPANY := '$(PRODUCTION_MACOS_BUNDLE_COMPANY)'
+	PRODUCTION_MACOS_BUNDLE_DISPLAY_NAME := '$(PRODUCTION_MACOS_BUNDLE_DISPLAY_NAME)'
+	PRODUCTION_MACOS_BUNDLE_NAME := '$(PRODUCTION_MACOS_BUNDLE_NAME)'
+	PRODUCTION_FOLDER_MACOS := $(PRODUCTION_FOLDER)
+	PRODUCTION_FOLDER := $(PRODUCTION_FOLDER)/$(PRODUCTION_MACOS_BUNDLE_NAME).app/Contents
+	PRODUCTION_FOLDER_RESOURCES := $(PRODUCTION_FOLDER)/Resources
+	PRODUCTION_DEPENDENCIES := $(PRODUCTION_DEPENDENCIES)
+	PRODUCTION_MACOS_DYLIBS := $(PRODUCTION_MACOS_DYLIBS:%=%.dylib)
+	MACOS_FRAMEWORKS?=CoreFoundation
+	PRODUCTION_MACOS_FRAMEWORKS := $(PRODUCTION_MACOS_FRAMEWORKS:%=%.framework)
+	PRODUCTION_MACOS_BACKGROUND := env/osx/$(PRODUCTION_MACOS_BACKGROUND)
+	MACOS_FRAMEWORK_PATHS := $(MACOS_FRAMEWORK_PATHS:%=-F%)
+	BUILD_FLAGS := $(BUILD_FLAGS) $(MACOS_FRAMEWORK_PATHS) $(MACOS_FRAMEWORKS:%=-framework %)
+endif
 
 #==============================================================================
 # Directories & Dependencies
 BLD_DIR := bin/$(BUILD)
+ifeq ($(BUILD),Tests)
+	BLD_DIR := bin/Release
+endif
 BLD_DIR := $(BLD_DIR:%/=%)
 TARGET := $(BLD_DIR)/$(NAME)
 _NAMENOEXT := $(NAME:.exe=)
 _NAMENOEXT := $(_NAMENOEXT:.dll=)
+
+ifneq ($(SRC_TARGET),)
+	LIB_DIRS := $(LIB_DIRS) $(BLD_DIR)
+endif
+_LIB_DIRS := $(LIB_DIR:%=-L%/) $(LIB_DIRS:%=-L%)
 
 _SOURCES_IF_RC := $(if $(filter windows,$(PLATFORM)),$(SOURCE_FILES:.rc=.res),$(SOURCE_FILES:%.rc=))
 
@@ -114,7 +175,9 @@ DEP_SUBDIRS := $(PROJECT_DIRS:%=$(DEP_DIR)/%)
 _PCH_HFILE := $(shell find $(SRC_DIR) -name '$(PRECOMPILED_HEADER).hpp' -o -name '$(PRECOMPILED_HEADER).h' -o -name '$(PRECOMPILED_HEADER).hh')
 _PCH_HFILE := $(_PCH_HFILE:$(SRC_DIR)/%=%)
 _PCH_EXT := $(_PCH_HFILE:$(PRECOMPILED_HEADER).%=%)
-_PCH_COMPILER_EXT := gch
+_PCH_COMPILER_EXT := $(if $(filter osx,$(PLATFORM)),p,g)ch
+
+_SYMBOLS := $(if $(filter osx,$(PLATFORM)),,$(if $(filter Release,$(BUILD)),-s,))
 
 
 _PCH := $(_PCH_HFILE:%=$(OBJ_DIR)/%)
@@ -154,7 +217,11 @@ OBJ_COMPILE = $(CC) $(CFLAGS_DEPS) $(_BUILD_MACROS) $(_INCLUDE_DIRS) $(_INCLUDE_
 OBJ_COMPILE_T = $(CC) $(CFLAGS_DEPS_T) $(_BUILD_MACROS) $(_INCLUDE_DIRS) $(_INCLUDE_PCH) $(CFLAGS) -o $@ -c $<
 
 RC_COMPILE = -$(RC) -J rc -O coff -i $< -o $@
-ASM_COMPILE = objdump -d -C -Mintel $< > $@
+ifeq ($(PLATFORM),osx)
+	ASM_COMPILE = otool -tvV $< | c++filt > $@
+else
+	ASM_COMPILE = objdump -d -C -Mintel $< > $@
+endif
 POST_COMPILE = mv -f $(DEP_DIR)/$*.Td $(DEP_DIR)/$*.d && touch $@
 POST_COMPILE_T = mv -f $(DEP_DIR)/.$(TEST_DIR)/$*.Td $(DEP_DIR)/.$(TEST_DIR)/$*.d && touch $@
 
@@ -173,13 +240,11 @@ buildprod: all makeproduction
 
 #==============================================================================
 # Functions
-define color_reset
-	@tput setaf 4
-endef
+color_reset := @tput setaf 4
 
-define comple_with
+define compile_with
 	$(color_reset)
-	$(if $(_CLEAN),@echo $($(2):$(OBJ_DIR)/%=%))
+	$(if $(_CLEAN),@echo '   $($(2):$(OBJ_DIR)/%=%)')
 	$(_Q)$(3) && $(4)
 endef
 
@@ -191,27 +256,31 @@ makepch: $(_PCH_GCH)
 
 makebuild: $(TARGET)
 	$(color_reset)
-	@echo '$(BUILD) build target is up to date.'
+ifeq ($(SRC_TARGET),)
+	@echo '   Target is up to date.'
+else
+	@echo '   $(NAME): Target is up to date.'
+endif
 .PHONY: makebuild
 
 #==============================================================================
 # Build Recipes
 $(OBJ_DIR)/%.o: $(SRC_DIR)/%
 $(OBJ_DIR)/%.o: $(SRC_DIR)/% $(_PCH_GCH) $(DEP_DIR)/%.d | $(_DIRECTORIES)
-	$(call comple_with,@,<,$(OBJ_COMPILE),$(POST_COMPILE))
+	$(call compile_with,@,<,$(OBJ_COMPILE),$(POST_COMPILE))
 
 $(OBJ_DIR)/.$(TEST_DIR)/%.o: $(TEST_DIR)/%
 $(OBJ_DIR)/.$(TEST_DIR)/%.o: $(TEST_DIR)/% $(_PCH_GCH) $(DEP_DIR)/.$(TEST_DIR)/%.d | $(_DIRECTORIES)
-	$(call comple_with,@,<,$(OBJ_COMPILE_T),$(POST_COMPILE_T))
+	$(call compile_with,@,<,$(OBJ_COMPILE_T),$(POST_COMPILE_T))
 
 $(OBJ_DIR)/%.$(_PCH_EXT).$(_PCH_COMPILER_EXT) : $(SRC_DIR)/%.$(_PCH_EXT)
 $(OBJ_DIR)/%.$(_PCH_EXT).$(_PCH_COMPILER_EXT) : $(SRC_DIR)/%.$(_PCH_EXT) $(DEP_DIR)/%.d | $(_DIRECTORIES)
-	$(call comple_with,@,<,$(PCH_COMPILE),$(POST_COMPILE))
+	$(call compile_with,@,<,$(PCH_COMPILE),$(POST_COMPILE))
 
 $(OBJ_DIR)/%.res: $(SRC_DIR)/%.rc
 $(OBJ_DIR)/%.res: $(SRC_DIR)/%.rc $(DEP_DIR)/%.d | $(_DIRECTORIES)
 	$(color_reset)
-	$(if $(_CLEAN),@echo $(<:$(OBJ_DIR)/%=%))
+	$(if $(_CLEAN),@echo "   $(<:$(OBJ_DIR)/%=%)")
 	$(_Q)$(RC_COMPILE)
 
 $(ASM_DIR)/%.o.asm: $(OBJ_DIR)/%.o
@@ -220,15 +289,23 @@ $(ASM_DIR)/%.o.asm: $(OBJ_DIR)/%.o
 
 $(TARGET): $(_PCH_GCH) $(OBJS) $(ASMS) $(TEST_DIR)
 	$(color_reset)
-	$(if $(_CLEAN),@echo; echo 'Linking: $(TARGET)')
+	$(if $(_CLEAN),@echo; printf '\xE2\x87\x9B'; echo '  Linking: $(TARGET)')
+ifeq ($(suffix $(TARGET)),.dll)
 ifeq ($(BUILD_STATIC),true)
 	-$(_Q)rm -rf $(BLD_DIR)/lib$(_NAMENOEXT).a
 	$(_Q)ar.exe -r -s $(BLD_DIR)/lib$(_NAMENOEXT).a $(OBJS)
 else
 	-$(_Q)rm -rf $(BLD_DIR)/lib$(_NAMENOEXT).def $(BLD_DIR)/lib$(_NAMENOEXT).a
-	$(_Q)$(CC) -shared -Wl,--output-def="$(BLD_DIR)/lib$(_NAMENOEXT).def" -Wl,--out-implib="$(BLD_DIR)/lib$(_NAMENOEXT).a" -Wl,--dll $(_LIB_DIRS) $(OBJS) -o $@ -s $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+	$(_Q)$(CC) -shared -Wl,--output-def="$(BLD_DIR)/lib$(_NAMENOEXT).def" -Wl,--out-implib="$(BLD_DIR)/lib$(_NAMENOEXT).a" -Wl,--dll $(_LIB_DIRS) $(OBJS) -o $@ $(_SYMBOLS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
 endif
-	$(foreach dep,$(BUILD_DEPENDENCIES),$(shell cp -r $(dep) $(BLD_DIR)))
+else
+	$(_Q)$(CC) $(_LIB_DIRS) $(_SYMBOLS) -o $@ $(OBJS) $(_LINK_LIBRARIES) $(BUILD_FLAGS)
+endif
+	@echo
+ifneq ($(BUILD_DEPENDENCIES),)
+	$(foreach dep,$(BUILD_DEPENDENCIES),$(call copy_to,$(dep),$(BLD_DIR)))
+	@echo
+endif
 
 $(_DIRECTORIES):
 	$(if $(_CLEAN),,$(color_reset))
@@ -236,7 +313,7 @@ $(_DIRECTORIES):
 
 clean:
 	$(color_reset)
-	$(if $(_CLEAN),@echo 'Cleaning old build files & folders...'; echo)
+	$(if $(_CLEAN),@echo '   Cleaning old build files & folders...'; echo)
 	$(_Q)$(RM) $(TARGET) $(DEPS) $(OBJS)
 .PHONY: clean
 
@@ -245,7 +322,10 @@ clean:
 
 rmprod:
 	$(color_reset)
-	-$(_Q)rm -rf $(PRODUCTION_FOLDER)
+	-$(_Q)rm -rf $(if $(filter osx,$(PLATFORM)),$(PRODUCTION_FOLDER_MACOS),$(PRODUCTION_FOLDER))
+ifeq ($(PLATFORM),linux)
+	-$(_Q)rm -rf ~/.local/share/applications/$(NAME).desktop
+endif
 .PHONY: rmprod
 
 mkdirprod:
@@ -253,17 +333,83 @@ mkdirprod:
 	$(MKDIR) $(PRODUCTION_FOLDER)
 .PHONY: mkdirprod
 
+define copy_to
+	@printf "\xE2\x9E\xA6"
+	@echo "  Copying \"$(1)\" to \"$(CURDIR)/$(2)\""
+	$(shell cp -r $(1) $(2))
+endef
+
 releasetoprod: $(TARGET)
 	$(color_reset)
+ifeq ($(PLATFORM),osx)
+	@echo '   Creating the MacOS application bundle...'
+	@echo
+	$(MKDIR) $(PRODUCTION_FOLDER)/Resources $(PRODUCTION_FOLDER)/Frameworks $(PRODUCTION_FOLDER)/MacOS
+ifeq ($(shell brew ls --versions makeicns),)
+	brew install makeicns
+	$(color_reset)
+endif
+	$(_Q)makeicns -in env/osx/$(PRODUCTION_MACOS_ICON).png -out $(PRODUCTION_FOLDER)/Resources/$(PRODUCTION_MACOS_ICON).icns
+	@echo
+	$(_Q)plutil -convert binary1 env/osx/Info.plist.json -o $(PRODUCTION_FOLDER)/Info.plist
+	$(_Q)plutil -replace CFBundleExecutable -string $(NAME) $(PRODUCTION_FOLDER)/Info.plist
+	$(_Q)plutil -replace CFBundleName -string $(PRODUCTION_MACOS_BUNDLE_NAME) $(PRODUCTION_FOLDER)/Info.plist
+	$(_Q)plutil -replace CFBundleIconFile -string $(PRODUCTION_MACOS_ICON) $(PRODUCTION_FOLDER)/Info.plist
+	$(_Q)plutil -replace CFBundleDisplayName -string "$(PRODUCTION_MACOS_BUNDLE_DISPLAY_NAME)" $(PRODUCTION_FOLDER)/Info.plist
+	$(_Q)plutil -replace CFBundleIdentifier -string com.$(PRODUCTION_MACOS_BUNDLE_DEVELOPER).$(PRODUCTION_MACOS_BUNDLE_NAME) $(PRODUCTION_FOLDER)/Info.plist
+	$(_Q)cp $(TARGET) $(PRODUCTION_FOLDER)/MacOS
+	$(_Q)chmod +x $(PRODUCTION_FOLDER)/MacOS/$(NAME)
+else ifeq ($(PLATFORM),linux)
 	$(_Q)cp $(TARGET) $(PRODUCTION_FOLDER)
+	$(_Q)cp env/linux/$(PRODUCTION_LINUX_ICON).png $(PRODUCTION_FOLDER)/$(PRODUCTION_LINUX_ICON).png
+	$(_Q)cp env/linux/exec.desktop $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)sed -i 's/^Exec=.*/Exec=$(_LINUX_GREP_CWD)\/$(PRODUCTION_FOLDER)\/$(NAME)/' $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)sed -i 's/^Path=.*/Path=$(_LINUX_GREP_CWD)\/$(PRODUCTION_FOLDER)/' $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)sed -i 's/^Name=.*/Name=$(PRODUCTION_LINUX_APP_NAME)/' $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)sed -i 's/^Comment=.*/Comment=$(PRODUCTION_LINUX_APP_COMMENT)/' $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)sed -i 's/^Icon=.*/Icon=$(_LINUX_GREP_CWD)\/$(PRODUCTION_FOLDER)\/$(PRODUCTION_LINUX_ICON).png/' $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)chmod +x $(PRODUCTION_FOLDER)/$(NAME)
+	$(_Q)chmod +x $(PRODUCTION_FOLDER)/$(NAME).desktop
+	$(_Q)cp $(PRODUCTION_FOLDER)/$(NAME).desktop ~/.local/share/applications
+else
+	$(_Q)cp $(TARGET) $(PRODUCTION_FOLDER)
+endif
 .PHONY: releasetoprod
 
 makeproduction: rmprod mkdirprod releasetoprod
 	$(color_reset)
-	@echo -n 'Adding dynamic libraries & project dependencies...'
-	$(foreach dep,$(PRODUCTION_DEPENDENCIES),$(shell cp -r $(dep) $(PRODUCTION_FOLDER_RESOURCES)))
+ifneq ($(PRODUCTION_DEPENDENCIES),)
+	@echo '   Adding dynamic libraries & project dependencies...'
+	@echo
+	$(foreach dep,$(PRODUCTION_DEPENDENCIES),$(call copy_to,$(dep),$(PRODUCTION_FOLDER_RESOURCES)))
 	$(foreach excl,$(PRODUCTION_EXCLUDE),$(shell find $(PRODUCTION_FOLDER_RESOURCES) -name '$(excl)' -delete))
-	@echo ' Done'
+endif
+ifeq ($(PLATFORM),osx)
+	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(call copy_to,$(dylib),$(PRODUCTION_FOLDER)/MacOS))
+	$(_Q)install_name_tool -add_rpath @executable_path/../Frameworks $(PRODUCTION_FOLDER)/MacOS/$(NAME)
+	$(_Q)install_name_tool -add_rpath @loader_path/.. $(PRODUCTION_FOLDER)/MacOS/$(NAME)
+	$(foreach dylib,$(PRODUCTION_MACOS_DYLIBS),$(shell install_name_tool -change $(notdir $(dylib)) @rpath/MacOS/$(notdir $(dylib)) $(PRODUCTION_FOLDER)/MacOS/$(NAME)))
+	$(foreach framework,$(PRODUCTION_MACOS_FRAMEWORKS),$(call copy_to,$(framework),$(PRODUCTION_FOLDER)/Frameworks))
+ifeq ($(PRODUCTION_MACOS_MAKE_DMG),true)
+	$(shell hdiutil detach /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/ &> /dev/null)
+	@echo
+	@echo '   Creating the dmg image for the application...'
+	@echo
+	$(_Q)hdiutil create -megabytes 54 -fs HFS+ -volname $(PRODUCTION_MACOS_BUNDLE_NAME) $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg > /dev/null
+	$(_Q)hdiutil attach $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg > /dev/null
+	$(_Q)cp -r $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).app /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/
+	-$(_Q)rm -rf /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/.fseventsd
+	$(MKDIR) /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/.background
+	$(_Q)tiffutil -cathidpicheck $(PRODUCTION_MACOS_BACKGROUND).png $(PRODUCTION_MACOS_BACKGROUND)@2x.png -out /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/.background/background.tiff
+	$(_Q)ln -s /Applications /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/Applications
+	$(_Q)appName=$(PRODUCTION_MACOS_BUNDLE_NAME) osascript env/osx/dmg.applescript
+	$(_Q)hdiutil detach /Volumes/$(PRODUCTION_MACOS_BUNDLE_NAME)/ > /dev/null
+	$(_Q)hdiutil convert $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg -format UDZO -o $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).dmg > /dev/null
+	$(_Q)rm -f $(PRODUCTION_FOLDER_MACOS)/.tmp.dmg
+	@echo
+	@echo '   Created $(PRODUCTION_FOLDER_MACOS)/$(PRODUCTION_MACOS_BUNDLE_NAME).dmg'
+endif
+endif
 .PHONY: makeproduction
 
 #==============================================================================
